backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==1.0.4
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
PyJWT==2.8.0